Private header files should not be installed.
Hiding them in the .cc file is the best way to prevent this abuse.
